Support install without Cython already installed
The primary functionality change is that pip install ... should now work even in a fresh environment without Cython installed.

As of setuptools==18.0, Cython extensions can be built with just the Cython dependency specified in setup_requires. The direct import at the top of setup.py was preventing this behaviour as the import fails and terminates the code even though the Cython setup dependency was specified.